backend/FunctionApp/requirements.txt
azure-functions==1.21.3
certifi==2025.1.31
charset-normalizer==3.4.1
idna==3.10
requests==2.32.3
urllib3==2.3.0
azure-storage-blob==12.10.0
